Some minor ux changes and fixes

+++ app.py
@@ -9,14 +9,23 @@ import requests
 import glob
 import time 
 import threading
+import logging
     
-
 app = Flask(__name__)
 main_api_token = createToken(10)
 socket = SocketIO(app)
 yt_video = "https://youtube.com/watch?v="
-downloads_folder = join("downloads")
-max_cache_time = 5400
+max_cache_time = 30
+
+tempCsettings = readJson("settings.json")
+if tempCsettings['downloads_folder'][0] == 'cwd':
+    tempCsettings['downloads_folder'][0] = os.getcwd()
+
+if tempCsettings['database'][0] == 'cwd':
+    tempCsettings['database'][0] = os.getcwd()
+
+downloads_folder = join(tempCsettings['downloads_folder'])
+database_path = tempCsettings['database'] # Bad name ik
 
 @socket.on("convertRequest")
 def on_convertRequest(data):
@@ -24,11 +33,33 @@ def on_convertRequest(data):
     download_api = root + "api/v1/download"
     requests.post(f"{download_api}?id={data['id']}&token={main_api_token}&sid={request.sid}")
 
+@socket.on("update_item_database")
+def update_item_database(data):
+    _id = data['video_id']
+    database = readJson(database_path)
+
+    for i, item in enumerate(database['data']):
+        if item['id'] == _id:
+            database['data'][i]['downloads'] += 1
+            database['data'][i]['timestamp'] = int(time.time())
+
+            downloads = database['data'][i]['downloads']
+            timestamp = database['data'][i]['timestamp']
+
+            socket.emit("update_results", {
+                "downloads": downloads,
+                "last_download": timestamp,
+                "id": _id
+            })
+
+            break
+    writeJson(database_path, database)
+
 @app.route("/api/v1/get_song")
 def get_song():
     _id = request.args.get("id")
     title = request.args.get("title")
-
+    
     return send_file(join(f"{downloads_folder}/{_id}.mp3"), attachment_filename=title + ".mp3", as_attachment=True)
 
 @app.route("/api/v1/download", methods=["POST"])
@@ -81,40 +112,47 @@ def download():
                         "thumbnail": info['thumbnail']
                     }, room=sid)
                     
-                    database = readJson(f"{downloads_folder}/database.json")
+                    database = readJson(database_path)
 
                     createData = True
                     cached = True
+                    appendToDatabase = True
 
-                    if glob.glob(join(f"{downloads_folder}/{info['id']}.*")):
-                        print("File exists")
-                        for i, item in enumerate(database['data']):
-                            if item['id'] == info['id']:
-                                database['data'][i]['downloads'] += 1
-                                database['data'][i]['timestamp'] = int(time.time())
-                                
-                                downloads = database['data'][i]['downloads']
+                    # if glob.glob(join(f"{downloads_folder}/{info['id']}.*")):
+                    for i, item in enumerate(database['data']):
+                        if item['id'] == info['id']:
+                            appendToDatabase = False
+
+                            downloads = database['data'][i]['downloads']
+                            timestamp = 0
+
+                            if item['cached']:
                                 timestamp = database['data'][i]['timestamp']
 
                                 createData = False
+                            else:
+                                database['data'][i]['cached'] = True
+                            break
 
                     if createData:
                         ydl.download([url])
-                        data = {
-                            "id": info['id'],
-                            "title": info['title'],
-                            "timestamp": int(time.time()),
-                            "downloads": 1,
-                            "cached": True
-                        }
 
-                        downloads = data['downloads']
-                        timestamp = data['timestamp']
                         cached = False
 
-                        database['data'].append(data)
+                        if appendToDatabase:
+                            data = {
+                                "id": info['id'],
+                                "title": info['title'],
+                                "timestamp": 0,
+                                "downloads": 0,
+                                "cached": True
+                            }
 
-                    writeJson(f"{downloads_folder}/database.json", database)
+                            downloads = data['downloads']
+                            timestamp = data['timestamp']
+                            database['data'].append(data)
+
+                    writeJson(database_path, database)
 
                     socket.emit("convert_complete", {
                         "thumbnail": info['thumbnail'],
@@ -126,7 +164,8 @@ def download():
                         "last_download": timestamp,
                         "upload_date": info['upload_date'],
                         "cached": cached,
-                        "download_url": request.url_root + f"api/v1/get_song?id={info['id']}&title={info['title']}"
+                        "download_url": request.url_root + f"api/v1/get_song?id={info['id']}&title={info['title']}",
+                        "video_id": video_id
                     }, room=sid)
 
                 except youtube_dl.utils.DownloadError:
@@ -147,22 +186,26 @@ def index():
 
 def handleCachedSongs():
     while True:
-        database = readJson(f"{downloads_folder}/database.json")
+        database = readJson(database_path)
+
+        if max_cache_time == 0:
+            break
 
         for i, item in enumerate(database['data']):
             now = int(time.time())
             item_time = item['timestamp']
-            seconds_passed = now - item_time
-            
-            if seconds_passed > max_cache_time:
-                song_path = join(f"{downloads_folder}/{item['id']}.mp3")
-                if os.path.exists(song_path):
-                    os.remove(song_path)
+            if item_time != 0:
+                seconds_passed = now - item_time
+                
+                if seconds_passed > max_cache_time:
+                    song_path = join(f"{downloads_folder}/{item['id']}.mp3")
+                    if os.path.exists(song_path):
+                        os.remove(song_path)
+                        database['data'][i]['cached'] = False
+                        database['data'][i]['timestamp'] = 0
 
-                database['data'][:] = [d for d in database['data'] if d.get("id") != item['id']]
-                writeJson(f"{downloads_folder}/database.json", database)
-
-                print(f"{item['id']} Exceeded max cache time, deleted.")
+                    # database['data'][:] = [d for d in database['data'] if d.get("id") != item['id']]
+                    writeJson(database_path, database)
 
         time.sleep(10)
